requirements.txt
# Framework y Servidor
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
psycopg2-binary==2.9.9

# Autenticación y Seguridad
python-jose[cryptography]==3.3.0
# CAMBIO AQUÍ: Usamos una versión compatible
passlib[bcrypt]==1.7.4

# Configuración y Utilidades
python-dotenv==1.0.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Migraciones de Base de Datos
alembic==1.13.1

# Testing
pytest==8.2.0
httpx==0.27.0
pytest-cov==5.0.0